Use dependency instead of find_library and avoid relying on rizin exec
pkg-config files already contain all the required include and library
dependencies (plus optional CFLAGS) necessary to compile against Rizin
libraries. One can specify --pkg-config-path to specify the pkg-config
path to use if not in the standard directory. Also, by setting the
plugdir to prefix/lib/rizin/plugins (instead of RZ_USER_PLUGINS from the
rizin executable) the person compiling the plugin can choose whether to
install the plugin in a system-wide location (e.g. a maintainer of a
linux distro who wants to ship jsdec as a package would want to install
it in /usr/lib64/rizin/plugins and not in ~/.local/lib64/rizin/plugins)
or in the home location.